+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * **Breaking Change:** cli arguments `--fatal-unused` and `--fatal-warnings` activate by default.
 * chore: restrict `analyzer` version to `>=2.8.0 <3.3.0`.
-* feat: add static code diagnostic `ban-name`
+* feat: add static code diagnostics `avoid-collection-methods-with-unrelated-types`, `ban-name`.
 
 ## 4.11.0